% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elec_results_map.R
\name{elec_winners_map}
\alias{elec_winners_map}
\title{elec_results_map
STILL IN PROGRESS! This function takes a vote results file that has been run through the elec_find_winner and elec_pcts_by_row functions and combined with a GIS shapefile. The result is a Leaflet map with diverging colors for the winning and losing candidates/positions, and an optional turnout layer. Vignette to come.}
\usage{
elec_winners_map(gisdatafile, winner = "Yes", loser = "No",
  placeCol = "Precinct", wincolors = "Greens", losecolors = "Reds",
  turnout = TRUE, turnoutcolors = "Greens", WinPctMargin = "YesPctMargin")
}
\arguments{
\item{gisdatafile}{- sf (simple features) geospatial file combined with vote results data that was already processed with elec_find_winner and elec_pcts_by_row}

\item{winner}{- string name of winning candidate or position}

\item{loser}{- string name of losing candidate or position}

\item{placeCol}{- string name of column that holds place names. If you want names to display as "Precinct 1" you must have a placeCol column that combines the string "Precinct " with the precinct number.}

\item{wincolors}{- string name of ColorBrewer palette. Defaults to "Greens."}

\item{losecolors}{- string name of ColorBrewer palette. Defaults to "Reds."}

\item{turnout}{- logical Whether data includes a column named Turnout with turnout percents as a fraction}

\item{turnoutcolors}{- string name of ColorBrewer palette. Defaults to "Greens."}

\item{WinPctMargin}{- string name of column with the difference between the winning percent and losing percent for each place. May need to be calculated manually.}
}
\value{
leaflet object
}
\description{
elec_results_map
STILL IN PROGRESS! This function takes a vote results file that has been run through the elec_find_winner and elec_pcts_by_row functions and combined with a GIS shapefile. The result is a Leaflet map with diverging colors for the winning and losing candidates/positions, and an optional turnout layer. Vignette to come.
}

